+++ CHANGELOG.md
@@ -16,5 +16,9 @@
 
 - **Menu background can now be enabled in all menus** instead of just in setup screens
 - **More menu background choices:** Solid (formerly "on"), Dark, Darker
-- **Improved HUD color customization**
+- **Improved HUD color customization:** disabling of the bars in Boom HUD, ability to choose colors for things like:
+	- *Widget names*
+	- *Plain text*
+	- *Total level time*
+	- *Low/Ok/Good/Extra thresholds* for Health, Armor and Ammo widgets
 - **Rearranged Status Bar/HUD and Automap setup menus**